+++ src/screens/index.ts
@@ -1,2 +1,4 @@
 export * from "./HomeScreen";
 export * from "./SettingsScreen";
+export * from "./LeaderboardScreen";
+export * from "./ProfileScreen";